refactor(electron): group services and ipc folders
