+++ electron/services/ProvidersService.ts
@@ -1,8 +1,8 @@
 import { and, asc, eq, sql } from 'drizzle-orm';
 import type { BetterSQLite3Database } from 'drizzle-orm/better-sqlite3';
-import * as schema from './db/schema.js';
-import { providers, providerModels } from './db/schema.js';
-import type { ModelCapability, ProviderModelRow, ProviderRow } from './db/schema.js';
+import * as schema from '../db/schema.js';
+import { providers, providerModels } from '../db/schema.js';
+import type { ModelCapability, ProviderModelRow, ProviderRow } from '../db/schema.js';
 import { SafeStorageService } from './SafeStorageService.js';
 
 export interface ModelDTO {
