+++ electron/services/__tests__/HotkeysService.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { makeTestDb } from '../db/__tests__/testDb.js';
+import { makeTestDb } from '../../db/__tests__/testDb.js';
 import { HotkeysService } from '../HotkeysService.js';
 
 function newService() {
