+++ electron/services/__tests__/ProvidersService.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { makeTestDb } from '../db/__tests__/testDb.js';
+import { makeTestDb } from '../../db/__tests__/testDb.js';
 import { ProvidersService } from '../ProvidersService.js';
 import { SafeStorageService } from '../SafeStorageService.js';
