+++ main.ts
@@ -12,26 +12,26 @@ import path from 'path';
 import { fileURLToPath } from 'url';
 import process from 'process';
 import { execFile } from 'child_process';
-import SettingsService from './electron/SettingsService';
-import SystemPermissionService from './electron/SystemPermissionService';
+import SettingsService from './electron/services/SettingsService';
+import SystemPermissionService from './electron/services/SystemPermissionService';
 import logService from './electron/LogService';
-import AutoUpdateService from './electron/AutoUpdateService';
-import OpenAtLoginService from './electron/OpenAtLoginService';
-import ThemeService from './electron/ThemeService';
-import pathService from './electron/PathService';
+import AutoUpdateService from './electron/services/AutoUpdateService';
+import OpenAtLoginService from './electron/services/OpenAtLoginService';
+import ThemeService from './electron/services/ThemeService';
+import pathService from './electron/services/PathService';
 import BetterSqlite3 from 'better-sqlite3';
 import { drizzle } from 'drizzle-orm/better-sqlite3';
 import { openDatabase } from './electron/db/client.js';
 import { resolveMigrationsFolder, runMigrations } from './electron/db/migrate.js';
 import { runImportIfNeeded } from './electron/db/import.js';
-import { ProvidersService } from './electron/ProvidersService.js';
-import { AgentsService } from './electron/AgentsService.js';
-import { HotkeysService } from './electron/HotkeysService.js';
+import { ProvidersService } from './electron/services/ProvidersService.js';
+import { AgentsService } from './electron/services/AgentsService.js';
+import { HotkeysService } from './electron/services/HotkeysService.js';
 import { runAgentImportIfNeeded } from './electron/db/importAgents.js';
-import type { HotkeyDTO } from './electron/HotkeysService.js';
+import type { HotkeyDTO } from './electron/services/HotkeysService.js';
 import * as dbSchema from './electron/db/schema.js';
 import { resolveUserDataPath } from './electron/userDataPath.js';
-import { registerIpcHandlers } from './electron/IpcHandlers.js';
+import { registerIpcHandlers } from './electron/ipc/registerIpc.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
